fix(mlanomalycharts): use int64 for max_series_to_plot to match its semantics
max_series_to_plot is a whole-number count of series, even though Kibana's
generated API type is a loosely-typed *float32. Expose it as an int64 in
the Terraform schema instead of float64/float32, converting at the API
boundary via typeutils.Int64ToFloat32Ptr / Float32PointerToInt64Pointer.

Adds two small reusable helpers for this int64<->float32 bridge:
typeutils.Float32PointerToInt64Pointer and panelkit.PreserveInt64.

+++ internal/kibana/dashboard/models/mlanomalycharts.go
@@ -23,7 +23,7 @@ import (
 
 type MlAnomalyChartsConfigModel struct {
 	JobIDs            []types.String                          `tfsdk:"job_ids"`
-	MaxSeriesToPlot   types.Float64                           `tfsdk:"max_series_to_plot"`
+	MaxSeriesToPlot   types.Int64                             `tfsdk:"max_series_to_plot"`
 	SeverityThreshold []MlAnomalyChartsSeverityThresholdModel `tfsdk:"severity_threshold"`
 	TimeRange         *TimeRangeModel                         `tfsdk:"time_range"`
 	Title             types.String                            `tfsdk:"title"`
